removed unnesscary classes
