--- src/main/java/frc/robot/commands/RaiseArm.java
@@ -1,20 +1,0 @@
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Arm;
-
-public class RaiseArm extends CommandBase{
-    
-    private Arm arm;
-
-    public RaiseArm(Arm arm) {
-        this.arm = arm;
-    }
-
-    @Override
-    public void execute() {
-        arm.raiseArm();
-    }
-
-}
-
