+++ src/main/java/frc/robot/commands/RunIntake.java
@@ -6,14 +6,21 @@ import frc.robot.subsystems.Intake;
 public class RunIntake extends CommandBase{
     
     private Intake intake;
-
-    public RunIntake(Intake intake) {
+    private boolean direction; //if true run intake forward, if false run intake backward
+    public RunIntake(Intake intake, boolean direction) {
         this.intake = intake;
+        this.direction = direction;
+        addRequirements(intake);
     }
 
     @Override
     public void execute() {
-        intake.runIntake();
+        if (direction == true) {
+            intake.runIntake();
+        }
+        else {
+            intake.runIntakeBackwards();
+        }
     }
