--- src/main/java/frc/robot/commands/RunIntakeBackwards.java
@@ -1,17 +1,0 @@
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Intake;
-
-public class RunIntakeBackwards extends CommandBase{
-    private Intake intake;
-
-    public RunIntakeBackwards(Intake intake) {
-        this.intake = intake;
-    }
-
-    @Override
-    public void execute() {
-        intake.runIntakeBackwards();
-    }
-}
